virt: keep on sampling before /var/run/vdsm/client.log exists
My commit 24c8e24 has introduced a serious bug: if vdsm is started
before /var/run/vdsm/client.log is created (which is the common case for
new installations), the host-sampling thread would stop when it fails to
find it.

This patch explicitly handles the case of a missing
/var/run/vdsm/client.log on the two locations that attempted to read its
time.

Change-Id: Ibb7d380882f5a93e5c75eda80598f382744b178c
Bug-Url: https://bugzilla.redhat.com/1111234
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/31619
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>