tool: Create configurators once
Commit e5e80540e8 replaced the configurators dict with a function,
creating a configuration dict on each call. The code added in this
commit invoke the function many times for each run on the tool, creating
10's of instances of the configurators. This change is not needed to
acomplish the purpuse of that patch, and is rather pointless.

This patch reverts the unrelated change, keeping the configurations in a
module constant. This change also simplify the tests.

Change-Id: I6e0e258dc28557661bf85734f000e43f20b1e512
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/31739
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32204
Tested-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: mooli tayer <mtayer@redhat.com>