tool: Make configurators more Pythonic
The configurations used to have Java like getters, which make the code
more clumsy then it should be. This patch convert the getters to read
only properties.

The types returned by the getters were mutable, which may lead to
calling code to modify the returned value without any error. Now the
configurators return immutable types.

There was no documentation about the purpose and the semantics of the
getters, which makes it harder for new developer to modify this part.
Now ModuleConfigure document these properties.

There should be no change in the behavior.

Change-Id: I16092da09d6763a8222bc941bd7d1501a7bb3bff
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/31741
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: mooli tayer <mtayer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>