tool: Remove duplicate names
The configurators modules were duplicating the module name in the class
name:

    libvirt.Libvrit

This is not only ugly, but does not describe what this class is. Now all
the configurtors modules define a Configurator class instead:

    libvirt.Configurator

This also has the nice property that it is now harder to import class
names into other modules, so people will import the modules instead.

Change-Id: I852019d5b5522a39c03dbf66a3f5383ccf584007
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/31775
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>