vm: declass warning for unsupported pause code
When VDSM starts, it resyncs the VM state with libvirt.
Unfortunately, libvirt does not provide yet all the
informations VDSM needs, most notably the pause reason
if a VM was paused.
Thus VDSM spews a warning about that.

We have bugs filed to the relevant dependencies since
a few months, so it's time to declass this warning
to a less worrysome message.

Change-Id: Id3cce078f1e357ce89c7352b2a88fc17ca1a2667
Related-To: https://bugzilla.redhat.com/1129587
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32154
Reviewed-by: Dan Kenigsberg <danken@redhat.com>