tc tests: Fix icmp test frame to have payload
In Ic6c5684ca8833807cf1dc05e9ccc4d97391c6168 I changed the original
packet to have a proper long and nice description. I generated the
frame with 'scapy' and I did not include any payload to ICMP.

For some reason, the jenkins unit tests pass but on my machines it
is causing the mirroring target not to see the frame. This patch
fixes the issue.

Change-Id: I5f2d64c3dbc621f062274afb343ad56048b36180
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32161
Reviewed-by: Dan Kenigsberg <danken@redhat.com>