ioprocess: Move and reword rename() warning
The warning was appearing even when people use the function in the
correct manner.

Let it be on the record that my original intent was to move the log line
to only display when a non-empty directory is being renamed instead of
every time.

But one of the requirements from my reviewers was to use the warnings
module instead making things more "correct" in the price of making the
line itself useless as no one turns on warnings in VDSM and there
is no way to atomically ignore warnings in a thread safe manner for
parts of the code so we can't even turn them on. Also, warnings go to
stderr by default which means that even if you turn them on no one will
see them.

Note: I only use the bug in the bug-url as it's partially related and it
feels to trivial to get a new bug opened and acked.

Change-Id: I6d83f91803515845c993c01c23390aaf1d757dff
Bug-Url: http://bugzilla.redhat.com/1131867
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32194
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>