getVMFullList fails when processing externally managed vms
When we send 22k response we get -1 during send. After the investigation
we found that m2crypto ignores error codes and returns -1. We are not
sure which error is returned. It can be one of SSL_ERROR_WANT_WRITE,
SSL_ERROR_WANT_READ, SSL_ERROR_NONE, SSL_ERROR_ZERO_RETURN.

When we get -1 we need to get back to poll loop so we override send
method and return 0 when we see -1 which fixes the issue.


Bug-Url: https://bugzilla.redhat.com/1135959
Change-Id: Ia728691880771388b6179e85a97df3f303aeddf8
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/32263
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>