getVMFullList hangs into infinite loop when handing many VMs
When we send 22k response we get -1 during send. After the investigation
we found that m2crypto ignores error codes and returns -1. We are not
sure which error is returned. It can be one of SSL_ERROR_WANT_WRITE,
SSL_ERROR_WANT_READ, SSL_ERROR_NONE, SSL_ERROR_ZERO_RETURN.

We use numSent in AsyncDispatcher#handle_write to decide whether we need
to process current buffer or we are done. When numSent is -1 we get into
infinite loop due to:

self._outbuf = data[numSent:]

We split the data and attempt to send it one more time but we still get
-1. We need to get to the poll loop to continue.


Reviewed-On: http://gerrit.ovirt.org/#/c/32263/
Bug-Url: https://bugzilla.redhat.com/1135959
Change-Id: Ia728691880771388b6179e85a97df3f303aeddf8
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/32290
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>