setNumberOfCpus marshalling issue
xmlrpc uses vmID twice to construct VM object and to run setNumberOfCpus.
In xmlrpc we provide implementation how we instantiate API object and
invoke a verb so we do not care how the parameters are used.

Whereas in jsonrpc we use "reflection" so we can't use the same parameter twice.


Bug-Url: https://bugzilla.redhat.com/1138181
Change-Id: Ie95c23c15bdb9b4932cd36b6f60bf84149c76454
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/32454
Reviewed-by: Dan Kenigsberg <danken@redhat.com>