virt: let Engine start a VM on an UNKOWN OS
This superflouse validation was introduced as a "bandage" to avoid
https://bugzilla.redhat.com/716705, where Vdsm failed to recognize
VMs with "UNKOWN" os injected to their bios.

Now that we recognize our VMs based on guest channels existence, we can
let Engine start a VM on whatever host it deems valid.

Backward compatibility caveat: if a VM is started on an UNKOWN os, and
then migrated to an old Vdsm (pre ovirt-3.4.2), it would be killed when
that old Vdsm is to be restarted. This would only affect users using
Fedora-based hosts with ovirt-3.3 and ovirt-3.6 on the same cluster.
Users of el-based hypervisors cannot migrate a VM started on a 3.6 host
to a 3.3 one since el7->el6 migration is blocked by Engine.

ovirt-3.3 has reached its end of supported life, so we should not care
about these purely-theoretical users.

Change-Id: Iee7b3913e76923043daab8fb85854c3290208237
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/32459
Reviewed-by: Francesco Romani <fromani@redhat.com>
Tested-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI