storage.monitor: Streamline names
Rename internal classes using shorter names, considering the package
namespace. For example:

    storage.monitor.DomainMonitorStatus -> storage.monitor.Status

In this case the name is more correct, as this is not the status of the
single DomainMonitor object but status of one of the monitor threads.

Since DomainMonitorThread was renamed, the logger was renamed too. We do
not need muliple loggers for the DomainMonitor and the MonitorThread, so
both of them use now the Storage.Monitor logger.

Change-Id: I8eae2f8548ebd1b336df10aea7dde47a4aeabf70
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32507
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>