ethtool_opts: provide a way to apply to all slaves
There was a hole in the feature that this hook services in that it
was not possible to apply some opts to all the slaves in the
cases when the combination of settings was not appliable to all
the slaves in a single ethtool command.

This patch solves the bug by adding a special device name that
the admin can use when defining the custom network opts, i.e.,
'*' that will mean that the special device name will be replaced
with each of the slave names, one at a time, and run as many ethtool
procs as there are slaves in the bond.

Bug-Url: https://bugzilla.redhat.com/1126478
Change-Id: Idddbc167be4ac8ba7c4b4ab10da0275b4caf2a58
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32784
Reviewed-by: Dan Kenigsberg <danken@redhat.com>