Ability to disable heartbeat
setupNetworks command reconfigures network so we may have temporary
disruption of heartbeats. in this situation we know that it may happen
so we need to ignore lack of heartbeats and restore it when we know that
we are done with network reconfiguration.

This patch provides api to update heartbeat policy. If the policy is
updated client reconnects with new settings.

Bug-Url: https://bugzilla.redhat.com/1136876
Change-Id: I9b7511422ab71a2e71091f69c70fbb97b04997c3
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>