tests: split tests fakes in a module
For the VM tests, we gathered a fair amount of Fake
helper objects. Is time to move them in a module,
in order to make it simpler for the other tests
to use them, and also to avoid to vmTests.py
become another vm.py, before it's too late.

Code changes are limited to renaming
* for better clarity; or
* to avoid redundancy or name clashing.

Change-Id: Ied7bb1d38d3814c406a53e1b7dbcac34eb6ffeb7
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32650
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>