config: introduce a devel section
This patch introduces a 'devel' configuration section,
in order to unclutter a bit the already crowded 'vars'
section.

An additional important benefit of this move is to clearly
mark that some tunables should never turned on in production
environments, unless exceptional circumstances, like a
performance problem being debugged.

For the same reason, is acceptable to move these options
across minor versions (e.g. 3.5 to 3.6).

This patch also makes room for future options which are
currently being discussed, like memory profiling or
a debug manhole service.

Change-Id: I7f9fc14c6fffd7ba111af985586eada83ec3c345
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32664
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>