unified_persistence: only consider bonds created/touched by vdsm
When upgrading a system to unified persistence we were considering
any non empty bond as a bond to persist under unified persistence.
This was not consistent with our implementation of ifcfg.flush that
only considers ifcfg files with the vdsm authoring header.

The patch fixes the behavior by making ifcfg flush and
unified_persistence upgrade consistent in it's ownership semantics,
i.e., we will not touch nor consider for persistence files not
markedly authored by vdsm.

A side effect of the lack of consistency was that a bond that was
providing connectivity to a machine could be made to be onboot no
if it was not immediately the backing device of ovirtmgmt.

Change-Id: Ic7f65546c55396ddf7022f5ffe17540ce0680d22
Bug-Url: https://bugzilla.redhat.com/1128140
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32703
Reviewed-by: Dan Kenigsberg <danken@redhat.com>