Do not add the memtune/min_guarantee element to the libvirt xml
VDSM does not need to put the min_guarantee to the libvirt XML.
It was only set and then nobody was accessing it. It will break
compatibility with tools that read data through libvirt directly,
but we do not officially support that anyway.

This part will fix all new VMs. Old VMs need a migration
hook that will be included as another patch.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1138340
Change-Id: Idb9a5a467ecee3235c842aa0b9b5fa9ccfd31ed9
Signed-off-by: Martin Sivak <msivak@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32774
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32788