hook: fix macspoof hook
The hook should be build with default VDSM build.

Enabling mac-spoofing is required for the appliance specifically for
dhcp agent which is connected to the networks bridge by a port (one or
more dhcp agent per network). In order for packets not to be blocked by
ebtable rules (introduced by nwfilter), mac-spoof should be enabled.

Change-Id: I924ba6e85f7299c72724c8b20a0e64d7b197ca23
Bug-Url: https://bugzilla.redhat.com/1139155
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32817
Reviewed-by: Dan Kenigsberg <danken@redhat.com>