virt: vm: sanitize earlier the 'smp' parameter
In the VM creation path, in the Vm._run()
method, we ensure the presence of the 'smp' parameter,
which is used in the creation path to build
the VM Domain XML to be fed to libvirt.

However, parameter sanitization is better done earlier,
so this patch moves the check in the Vm.__init__ method,
and also adds a couple of tests to ensure the 'smp'
parameter is correct.

Change-Id: I1b90a26ff6af560359f3d60a73afc56fb25dc15c
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32827
Reviewed-by: Dan Kenigsberg <danken@redhat.com>