vm: rename obsolete buildCmdLine method
In past times, VDSM used to run QEMU.
The Vm._buildCmdLine method was in charge
to build the QEMU executable command line.

Nowadays we use libvirt as middleware, with no
plans to come back, so this method does a different
thing.

This patch renames this method, with no behaviour
change, to reflect the reality.

Please note that this method is mostly orchestration
of vmxml.Domain calls;
Future patches will split this code from vm.py
to vmxml.py.

Change-Id: If4129948e71bf2308cce0b02f0d6c800903004bf
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32829
Reviewed-by: Dan Kenigsberg <danken@redhat.com>