qemuimg: Create qcow2 compat 0.10 images
Recent qemu versions introduced a new default qcow2 format (1.1), which
is not compatible with older versions of qemu. For example, EL6 host
cannot create snapshots if an image was created on EL7 host.

This patch uses the "compat" option to create images in the older format
(0.10), which are are compatible with different versions of qemu.

The "compat" option is not available on older qemu-img, so we have to
use the "-o ?" option to detect the availability of the "compat" option
before creating an image. This is the same method used by libvirt.

When all supported platforms provide the "compat" option, we can remove
the "compat" detection code. Chanding the compat version will require
a new storage domain format.

Change-Id: Iffd45b394c49e8b12fb7a4cbaa5c7a3519a2cc1c
Bug-Url: https://bugzilla.redhat.com/1139707
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32836
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33020
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>