Configure iSCSI iface.net_ifacename
When using multiple local devices in same subnet to have MPIO, we need to bind
each iscsi session to a network device, otherwise system will choose any
interface in the same subnet to use for the iSCSI session.

Engine now sends the netIfaceName, which defines the network device to be used
for each iscsi session. This patch enables vdsm to consume netIfaceName and
configures iscsi iface accordingly.

Change-Id: I3ebb2f272669b753700b57486d869b21dd16f2d6
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1112861
Signed-off-by: Amador Pahim <apahim@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/31534
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>