vm: implement VM.sdIds as read-only property
Changed the VM object sdIds list attribute. It used to be a list that
kept a copy of the storage domains in use in an additional list (with
all the difficulties of keeping the list in sync). That list has been
replaced with a property that returns the storage domains in use
getting them directly from the the vm devices in use.

Change-Id: Ie445d4689aa562b00229844007ffa05eaecebdcb
Relates-To: https://bugzilla.redhat.com/1094518
Signed-off-by: Fabio Pliger <fabio.pliger@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/32796
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>