Add a libvirt migration hook to filter out min_guarantee element
libvirt removed the silent support for min_guarantee element.
That will break migrations of long running machines and
resume of machines that were paused with the element still
in their XML.

VDSM was updated and new machines won't contain the offending
element. This fix will remove it when a running machine is migrated
or resumed.

Change-Id: I5f5e79037b755b257f9cac419887606f66705504
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1138340
Signed-off-by: Martin Sivak <msivak@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/32928
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Francesco Romani <fromani@redhat.com>