virt: migration: rename cancel->stop method
The most common meaning of  'cancel' is
'halt operation and discard the current progress',
which DowntimeThread doesn't do -and there is
no point in doing.

What cancel() method does is actually just halt
the thread without affecting the output, so this
patch renames the method to stop() in order
to make it less surprising and and also consistent
with migration.MonitorThread.

Besides that there are no changes in behaviour.

Change-Id: Ia938ae0fa5bf3ddd16ffc992299566b618ea4677
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33030
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>