gluster: Temporary fix for supervdsm memory leak.
     Supervdsm memory leak is caused by memory leak in libgfapi
(BZ:1093594) which is used by verb glusterVolumeStatsInfoGet to
retrive the volume statistics. The libgfapi method glfs_fini()
is not cleaning up the memory, so everytime the verb is invoked
memory gets added up(~20-30MiB). This verb is invoked by engine
every 5 minutes for volume capacity feature. So supervdsm memory
usage shoots up when gluster is enabled.

    This patch provides a temporary fix until libgfapi issue is
resolved. In this patch libgfapi method calls are executed as
script using execCmd(), by this libgfapi invocation is not part
of supervdsm process.

Change-Id: I7079426178ce47008d9a3b83635afce98536ca34
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1142647
Signed-off-by: ndarshan <dnarayan@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33312
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33950