netinfo, tool: Add a vdsm-tool to determine bonding defaults offline
dump-bonding-defaults reads bonding option defaults (per mode) and saves them
to /var/lib/vdsm/bonding-defaults.json. VDSM loads the file once at runtime
and refers to this static information of defaults when reporting bonding
options of bonds in the system.

As the bonding defaults are bound to the used kernel it is the responsibility
of the packager (or the sysadmin in customized setups) to regenerate the file
by running (during the build, or installation, respectively):

vdsm-tool dump-bonding-defaults

It is currently not advised to run the tool on startup because NetworkManager
may ifup the reference bond, making the tool unable to determine all options.

Change-Id: I89b907ba80f23f417d5e481db9350247445ab772
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1142701
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33403
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>