tool: remove AUTHOR from vdsm-tool manual page.
"lists authors of the documentation or program.  Use of an AUTHORS
section is strongly discouraged.  Generally, it is better not  to
clutter  every  page with a list of (over time potentially numer‐
ous) authors; if you write or significantly amend a page,  add  a
copyright notice as a comment in the source file.  If you are the
author of a device driver and want  to  include  an  address  for
reporting bugs, place this under the BUGS section."

see "man 7 man-pages" from the man-pages package .

Change-Id: I73a0fadec63c6643ae8d3678b611bbf15710142a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1146469
Signed-off-by: Mooli Tayer <mtayer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33455
Reviewed-by: Dan Kenigsberg <danken@redhat.com>