vdsm: making PM 'on' and 'off' sync
Consider the following scenario:

DC1 with H1 (with PM) and H2 on cluster C1
another host H3 on DC1 cluster C2

When we are blocking the communication with iptables from H2 to H1 PM
card and use the default proxy preferences (cluster, dc) a Restart
operation will always fail.

VDSM should perform start/stop sync and return the correct script
returned code in order that engine will know that H2 fails to perform
the operation and will try to use H3 as a proxy for the failed operation

This patch changes the fenceNode function to be sync when 'on' 'off'
operation are used in the same manner of 'status' operation and return
the correct exit code of the invoked fencing script to the caller.

Change-Id: I81d45977664c99c954c4663a26642e6e21cc65e3
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1139643
Signed-off-by: Eli Mesika <emesika@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33469
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>