virt: sampling: more cautious disk stats check
In commit d6d0eb763e623 we changed the disk stats
collection to deal with the benign case of
known-missing disk stats samples, to avoid add
bogus error on the logs.

Unfortunately a case was missing in the live
storage migration flow, which causes

TypeError: argument of type 'NoneType' is not iterable

in the logs defeating the purposes of the
originating change.
This patch fixes this issue by adding the missing
check.

Change-Id: I059b69c33d45950f8377597ee8c6e7824e1ec223
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33482
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>