vdsm_reg: Make vdsm-reg persist the management bridge it creates
In the vdsm_reg flow, for the oVirt node, the management bridge is
created using addNetworks. This puts the management network in the
RunningConfig ( /var/run/vdsm/netconf/nets ). Then it notifies the
engine of the registration request.

When the engine approves the request, it starts the deployment, part
of which is to start vdsmd. In the common init tasks of vdsmd there is
the restore_nets task that, if /var/run/vdsm/nets_restored does not
exist, it will clear all the networks in RunningConfig and configure
again those in PersistentConfig. The issue is that in a clean
installation there won't be any PersistentConfig, and thus the system
will lose connectivity.

In order to solve this issue, the best course of action is to persist
RunningConfig as PersistentConfig as soon as the management network
has been created by vdsm-reg (and strictly before the registration
request is sent to the engine).

Change-Id: I9d35da853df485a9196ad0657f832b77ae65d24e
Bug-Url: https://bugzilla.redhat.com/1144639
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33411
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Tested-by: Douglas Schilling Landgraf <dougsland@redhat.com>