addNetwork script: Prevent empty bond and vlan from reaching RunningConfig
The consumers of the addNetwork script pass empty parameters, e.g. ""
in order to fill the positional arguments that the script takes. The
problem with that is that those parameters were being passed to
addNetwork without modification, and that includes the
_alterNetworkConfig wrapper that makes addNetwork requests reach
runningConfig (and eventually PersistentConfig).

Due to the issue above, a stored network could contain entries like
'bonding': '' and 'vlan': '' that would mess with selective network
restoration.

Change-Id: If6d56eefc05cdb7456f80b7ec13d0be8ad087aa3
Bug-Url: https://bugzilla.redhat.com/1144639
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33510
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Tested-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>