vm: log exceptional monitor unresponsive case
In past ages we added an explicit case to mark
monitor unresponsive.

This case may be no long relevant. In order
to gather evidence against or supporting
this hypotesis, add an explicit log line
when this case is applied.

Also, extract the relevant code in a helper
method, and add a couple of unit tests.

Change-Id: I04b88513d319d3f4530aa6d9f1b4a1c783b33347
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33870
Reviewed-by: Dan Kenigsberg <danken@redhat.com>