caps: Report online logical CPUs
Introduces a new information in the capabilities command which reports
the list of logical CPUs that are online during its execution.

This is necessary because the kernel allows logical CPUs to be disabled
and enable on-the-fly, and ppc64 hosts have only one enabled logical
CPU per core and several CPU threads disabled. So this patch is needed
in order to validate CPU pinning configurations and count the number of
logical CPUs properly.

Change-Id: Ia41df1758e629dc1087416c1dad3fa4d115243fb
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1148092
Signed-off-by: Vitor de Lima <vdelima@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33789
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>